Conflicts:
	VERSION

+++ dev/cosbench-s3/src/com/intel/cosbench/api/S3Stor/S3Storage.java
@@ -4,6 +4,8 @@ import static com.intel.cosbench.client.S3Stor.S3Constants.*;
 
 import java.io.*;
 
+import org.apache.http.HttpStatus;
+
 import com.amazonaws.*;
 import com.amazonaws.auth.*;
 import com.amazonaws.services.s3.*;
@@ -34,6 +36,8 @@ public class S3Storage extends NoneStorage {
     	endpoint = config.get(ENDPOINT_KEY, ENDPOINT_DEFAULT);
         accessKey = config.get(AUTH_USERNAME_KEY, AUTH_USERNAME_DEFAULT);
         secretKey = config.get(AUTH_PASSWORD_KEY, AUTH_PASSWORD_DEFAULT);
+
+        boolean pathStyleAccess = config.getBoolean(PATH_STYLE_ACCESS_KEY, PATH_STYLE_ACCESS_DEFAULT);
         
 		String proxyHost = config.get(PROXY_HOST_KEY, "");
 		String proxyPort = config.get(PROXY_PORT_KEY, "");
@@ -41,6 +45,7 @@ public class S3Storage extends NoneStorage {
         parms.put(ENDPOINT_KEY, endpoint);
     	parms.put(AUTH_USERNAME_KEY, accessKey);
     	parms.put(AUTH_PASSWORD_KEY, secretKey);
+    	parms.put(PATH_STYLE_ACCESS_KEY, pathStyleAccess);
     	parms.put(PROXY_HOST_KEY, proxyHost);
     	parms.put(PROXY_PORT_KEY, proxyPort);
 
@@ -57,6 +62,7 @@ public class S3Storage extends NoneStorage {
         AWSCredentials myCredentials = new BasicAWSCredentials(accessKey, secretKey);
         client = new AmazonS3Client(myCredentials, clientConf);
         client.setEndpoint(endpoint);
+        client.setS3ClientOptions(new S3ClientOptions().withPathStyleAccess(pathStyleAccess));
         
         logger.debug("S3 client has been initialized");
     }
@@ -97,8 +103,10 @@ public class S3Storage extends NoneStorage {
     public void createContainer(String container, Config config) {
         super.createContainer(container, config);
         try {
-        	logger.info("Creating " + container);
-            client.createBucket(container);
+        	if(!client.doesBucketExist(container)) {
+	        	logger.info("Creating " + container);
+	            client.createBucket(container);
+        	}
         } catch (Exception e) {
             throw new StorageException(e);
         }
@@ -124,8 +132,14 @@ public class S3Storage extends NoneStorage {
     public void deleteContainer(String container, Config config) {
         super.deleteContainer(container, config);
         try {
-        	logger.info("Deleting " + container);
-            client.deleteBucket(container);
+        	if(client.doesBucketExist(container)) {
+        		logger.info("Deleting " + container);
+        		client.deleteBucket(container);
+        	}
+        } catch(AmazonS3Exception awse) {
+        	if(awse.getStatusCode() != HttpStatus.SC_NOT_FOUND) {
+        		throw new StorageException(awse);
+        	}
         } catch (Exception e) {
             throw new StorageException(e);
         }
@@ -137,6 +151,10 @@ public class S3Storage extends NoneStorage {
         try {
         	logger.info("Deleting " + container + "\\" + object);
             client.deleteObject(container, object);
+        } catch(AmazonS3Exception awse) {
+        	if(awse.getStatusCode() != HttpStatus.SC_NOT_FOUND) {
+        		throw new StorageException(awse);
+        	}
         } catch (Exception e) {
             throw new StorageException(e);
         }
